Import 2.1.101

+++ drivers/block/raid0.c
@@ -104,7 +104,7 @@ static int raid0_run (int minor, struct md_dev *mddev)
     md_size[minor]/data->smallest->size +
     (md_size[minor]%data->smallest->size ? 1 : 0);
 
-  printk ("raid0 : Allocating %d bytes for hash.\n",sizeof(struct raid0_hash)*nb_zone);
+  printk ("raid0 : Allocating %ld bytes for hash.\n",(long)sizeof(struct raid0_hash)*nb_zone);
   if ((data->hash_table=vmalloc (sizeof (struct raid0_hash)*nb_zone)) == NULL)
   {
     vfree(data->strip_zone);
